{AH} fix a \0 issue in header, refactor AlignmentHeader to use classmethod factory methods
